fix: add work-around for logging bug when upgrading on windows
